feat: Add AlbumRepository class
This commit adds the `AlbumRepository` class to the `albums` domain. This class provides methods for posting and retrieving album data. The `postAlbum` method is used to create a new album by accepting a `postAlbumEntity` object and throwing an error if the method is not implemented. The `getAlbumById` method is used to retrieve an album by its ID, also throwing an error if the method is not implemented. This class serves as the repository for handling album data in the application.