feat: Implement PostAlbumUseCase
This commit adds the `PostAlbumUseCase` class to the `applications/use_case` directory. The `PostAlbumUseCase` class is responsible for executing the use case of posting an album. It takes a payload object, creates a `PostAlbum` entity, and calls the `postAlbum` method of the `albumRepository` to store the album data. The result is then transformed into a `PostedAlbum` entity and returned.

This use case provides the functionality to create and store albums in the application.