feat: Add GetAlbumByIdUseCase class
This commit adds the `GetAlbumByIdUseCase` class to the `applications/use_case` directory. The `GetAlbumByIdUseCase` class is responsible for executing the use case of retrieving an album by its ID. It takes a payload object with an `albumId` property, calls the `getAlbumById` method of the `albumRepository` to fetch the album data, and returns a `GetAlbum` entity.

This use case provides the functionality to retrieve album data by ID in the application.