feat: Add PutAlbum entity class
This commit adds the `PutAlbum` entity class to the `albums` domain. The `PutAlbum` class represents an album object with `name` and `year` properties. It includes a constructor that validates the payload and assigns the values to the corresponding properties. This entity class is used for updating album data in the application.