feat: Add PutAlbumByIdUseCase class
This commit adds the `PutAlbumByIdUseCase` class to the `applications/use_case` directory. The `PutAlbumByIdUseCase` class is responsible for executing the use case of updating an album by its ID. It verifies the existence of the album, creates a `PutAlbum` entity with the updated data, and calls the `putAlbumById` method of the `albumRepository` to update the album. This use case provides the functionality to update album data in the application.