feat: Add DeleteAlbumByIdUseCase class
This commit adds the `DeleteAlbumByIdUseCase` class to the `applications/use_case` directory. The `DeleteAlbumByIdUseCase` class is responsible for executing the use case of deleting an album by its ID. It verifies the existence of the album, and if found, calls the `deleteAlbumById` method of the `albumRepository` to delete the album. This use case provides the functionality to delete an album in the application.