feat: Add deleteAlbumById method to AlbumRepository
This commit adds the `deleteAlbumById` method to the `AlbumRepository` class in the `src/domains/albums/AlbumRepository.js` file. The method is responsible for deleting an album by its ID. It throws an error with the message "ALBUM_REPOSITORY.METHOD_NOT_IMPLEMENTED" to indicate that the method is not yet implemented.

This code change provides the foundation for implementing the functionality to delete an album in the application.