feat: Add DeleteAlbumByIdUseCase and deleteAlbumById method
This commit adds the `DeleteAlbumByIdUseCase` class to the `applications/use_case` directory and the `deleteAlbumById` method to the `AlbumRepository` class. The `DeleteAlbumByIdUseCase` class is responsible for executing the use case of deleting an album by its ID, while the `deleteAlbumById` method handles the actual deletion of the album. These changes provide the functionality to delete an album in the application.