feat: Add PostSongUseCase for posting a song
This commit adds the `PostSongUseCase` class to the `applications/use_case` directory. The `PostSongUseCase` class is responsible for executing the use case of posting a song. It creates a new `PostSong` entity based on the provided payload, calls the `postSong` method of the `songRepository` to post the song, and returns a `PostedSong` entity representing the posted song.

These code changes enable the functionality to post a song in the application.