feat: Add GetSongsUseCase for retrieving songs
This commit adds the `GetSongsUseCase` class to the `src/applications/use_case/GetSongsUseCase.js` file. The use case is responsible for retrieving songs from the song repository and returning them as instances of the `GetSongs` entity. This functionality enables the application to fetch songs and use them for further processing.