refactor: Update GetAlbum and GetSongs entities to use GetSongsItems
This commit refactors the `GetAlbum` and `GetSongs` entities to use the `GetSongsItems` class instead of the `GetSongs` class. The `GetAlbum` entity now checks if each song in the `songs` array is an instance of `GetSongsItems` instead of `GetSongs`. This change ensures that the songs in the album meet the data type specification.